tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.object.clamp.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.object.clamp</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.clamp</h1>
<p class="desc">ImageData.data clamps numbers to [0, 255]</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData.data clamps numbers to [0, 255]");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);

imgdata.data[0] = 100;
imgdata.data[0] = 300;
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");
imgdata.data[0] = 100;
imgdata.data[0] = -100;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");

imgdata.data[0] = 100;
imgdata.data[0] = 200+Math.pow(2, 32);
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");
imgdata.data[0] = 100;
imgdata.data[0] = -200-Math.pow(2, 32);
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");

imgdata.data[0] = 100;
imgdata.data[0] = Math.pow(10, 39);
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");
imgdata.data[0] = 100;
imgdata.data[0] = -Math.pow(10, 39);
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");

imgdata.data[0] = 100;
imgdata.data[0] = -Infinity;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 100;
imgdata.data[0] = Infinity;
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");


});
</script>
